backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# AI Clients
anthropic==0.42.0
openai==1.12.0
mistralai==1.0.0
httpx==0.27.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# File Processing
PyMuPDF==1.23.8
python-docx==1.1.0
Pillow==10.2.0

# Email
sendgrid==6.11.0

# Utilities
python-dotenv==1.0.0
pydantic[email]==2.8.2
pydantic-settings==2.4.0
aiofiles==23.2.1

# PDF Generation
api2pdf==2.1.0
markdown==3.5.2
